refactor: extract invitation sub-components into separate files
Move InvitationInfoText and InvitationActions out of
AcceptInvitationView.tsx into their own files.
